Relase docs v0.6.0 (#86)
* updated docs for release

* edit path for tests

* edit path for tests

* added pipeline to menu

* updated all files from haystack

* switch documents download to haystack master

+++ src/pages/benchmarks/versions/update_benchmarks.sh
@@ -1,14 +1,25 @@
 #!/bin/bash
 
+# Used branch
+echo "Haystack branch: $1"
+
 # Master data
-versions=("latest/" "v0.4.0/" "v0.5.0/") 
+versions=("latest/" "v0.4.0/" "v0.5.0/" "v0.6.0/") 
 
 for i in "${versions[@]}"
 do
-    rm ./src/pages/benchmarks/versions/${i}site/en/performance/*
-    rm ./src/pages/benchmarks/versions/${i}site/en/map/*
-    rm ./src/pages/benchmarks/versions/${i}site/en/speed/*
+    if [ -d "./src/pages/benchmarks/versions/${i}" ]; then
+        rm ./src/pages/benchmarks/versions/${i}site/en/performance/*
+        rm ./src/pages/benchmarks/versions/${i}site/en/map/*
+        rm ./src/pages/benchmarks/versions/${i}site/en/speed/*
+    fi
 done
+
+if [ ! -d "./src/pages/benchmarks/versions/${i}" ]; then
+    mkdir -p ./src/pages/benchmarks/versions/${i}site/en/performance
+    mkdir -p ./src/pages/benchmarks/versions/${i}site/en/map
+    mkdir -p ./src/pages/benchmarks/versions/${i}site/en/speed
+fi
 
 for i in "${versions[@]}"
 do
@@ -17,8 +28,8 @@ do
         j=''
     fi
 
-    wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/benchmarks/reader_performance.json -O ./src/pages/benchmarks/versions/${i}site/en/performance/reader_performance.json 
-    wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/benchmarks/retriever_performance.json -O ./src/pages/benchmarks/versions/${i}site/en/performance/retriever_performance.json 
-    wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/benchmarks/retriever_map.json -O ./src/pages/benchmarks/versions/${i}site/en/map/retriever_map.json 
-    wget https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/${j}_src/benchmarks/retriever_speed.json -O ./src/pages/benchmarks/versions/${i}site/en/speed/retriever_speed.json
+    wget https://raw.githubusercontent.com/deepset-ai/haystack/${1}/docs/${j}_src/benchmarks/reader_performance.json -O ./src/pages/benchmarks/versions/${i}site/en/performance/reader_performance.json 
+    wget https://raw.githubusercontent.com/deepset-ai/haystack/${1}/docs/${j}_src/benchmarks/retriever_performance.json -O ./src/pages/benchmarks/versions/${i}site/en/performance/retriever_performance.json 
+    wget https://raw.githubusercontent.com/deepset-ai/haystack/${1}/docs/${j}_src/benchmarks/retriever_map.json -O ./src/pages/benchmarks/versions/${i}site/en/map/retriever_map.json 
+    wget https://raw.githubusercontent.com/deepset-ai/haystack/${1}/docs/${j}_src/benchmarks/retriever_speed.json -O ./src/pages/benchmarks/versions/${i}site/en/speed/retriever_speed.json
 done
